Usa os.path.join para concatenar os arquivos, cria um arquivo de configuracao para DTD e afins, exclui alguns códigos em desuso, inclui versions.ini no setup.py
