+++ src/scielo/bin/xml/prodtools/settings/dtd_and_xsl/versions.ini
@@ -1,0 +1,33 @@
+[1.1]
+dtd_id = -//NLM//DTD JATS (Z39.96) Journal Publishing DTD v1.1 20151215//EN
+local = JATS-journalpublishing1.dtd
+remote = https://jats.nlm.nih.gov/publishing/1.1/JATS-journalpublishing1.dtd
+folder = j1.1
+dtd_path = j1.1/JATS-Publishing-1-1-MathML2-DTD/JATS-Publishing-1-1-MathML2-DTD
+xsl_prep_report = j1.1/xsl/nlm-style-5.25/nlm-stylechecker.xsl
+xsl_report = j1.1/xsl/nlm-style-5.25/style-reporter.xsl
+xsl_output_scielo = j1.1/xsl/sgml2xml/xml2pmc.xsl
+xsl_output_pmc = j1.1/xsl/sgml2xml/pmc.xsl
+
+[3.0]
+dtd_id = -//NLM//DTD Journal Publishing DTD v3.0 20080202//EN
+local = journalpublishing3.dtd
+remote = https://dtd.nlm.nih.gov/publishing/3.0/journalpublishing3.dtd
+dtd_path = v3.0/dtd
+xsl_prep_report = v3.0/xsl/nlm-style-4.6.6/nlm-stylechecker.xsl
+folder = v3.0
+xsl_report = v3.0/xsl/nlm-style-4.6.6/style-reporter.xsl
+xsl_output_scielo = v3.0/xsl/sgml2xml/xml2pmc.xsl
+xsl_output_pmc = v3.0/xsl/sgml2xml/pmc.xsl
+
+[1.0]
+dtd id = -//NLM//DTD JATS (Z39.96) Journal Publishing DTD v1.0 20120330//EN
+local = JATS-journalpublishing1.dtd
+remote = https://jats.nlm.nih.gov/publishing/1.0/JATS-journalpublishing1.dtd
+dtd_path = j1.0/dtd/jats1.0
+folder = j1.0
+xsl_prep_report = j1.0/xsl/nlm-style-5.13/nlm-stylechecker.xsl
+xsl_report = j1.0/xsl/nlm-style-5.13/style-reporter.xsl
+xsl_output_pmc = j1.0/xsl/sgml2xml/pmc.xsl
+xsl_output_scielo = j1.0/xsl/sgml2xml/xml2pmc.xsl
+
