+++ src/scielo/bin/xml/setup.py
@@ -61,6 +61,7 @@ setup(
                       'settings/*/*.txt',
                       'settings/*/*.csv',
                       'settings/*/*.fst',
+                      'settings/*.ini',
                       ],
     },
     include_package_data=True,
